requirements.txt
# === Flask Backend ===
flask==2.3.3
werkzeug==2.3.7

# === CNN Classification (TensorFlow + EfficientNetB0) ===
tensorflow==2.13.0
numpy==1.23.5
Pillow==10.2.0

# === Real-ESRGAN Enhancement ===
# For CPU-only (safe default):
torch==2.0.1+cpu
# If using GPU with CUDA 11.8, uncomment below and comment above:
# torch==2.0.1+cu118

realesrgan
basicsr

# === YOLOv8 Stage Detection ===
ultralytics==8.1.20
opencv-python==4.8.0.76

# === Visualization & Evaluation ===
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0